Fix error when an absolute path is given
When an absolute path is given (e.g. via `lint-staged`), `node-ignore`
was throwing a RangeError `path should be a path.relative()'d string`.
see https://github.com/kaelzhang/node-ignore#1-pathname-should-be-a-pathrelatived-pathname

This is only a "quick win" to fix this issue.
As `prepareFileList` was already returning an `absolute` property based
on the cwd, this commit only introduce a new `relative` property, which
is also based on the cwd.
No new specific test was written for this use case.